Move utils to sub directory.
